+++ src/ui/service/miniquad/service.rs
@@ -1,21 +1,33 @@
 // devela::ui::service::miniquad::service
 //
-//! Defines [`MiniquadService`] and [`MiniquadEventHandlerExt`].
+//! Defines [`MiniquadEventHandlerExt`] and [`MiniquadService`].
 //
 
-#[cfg(feature = "alloc")]
-use crate::{Box, ToString};
 use crate::{
-    MiniquadConf, MiniquadEventHandler, UiCap, UiCapImage, UiCapInput, UiCapWindow, UiService,
+    iif, MiniquadConf, MiniquadEventHandler, UiCap, UiCapImage, UiCapInput, UiCapWindow, UiService,
 };
+use crate::{Box, ToString};
 #[cfg(doc)]
 use ::miniquad::FilterMode;
 
 /// A wrapper over `MiniquadEventHandler` to allow lazy initialization.
+///
+/// It also defines extra methods for convenience.
 pub trait MiniquadEventHandlerExt: MiniquadEventHandler {
     /// Returns the event handler initialized.
     fn init(self) -> Self;
+
+    /// Whether it interpolates the scaled pixels.
+    fn interpolation(&self) -> bool;
+    /// Set whether to `interpolate` the scaled pixels.
+    fn set_interpolation(&mut self, interpolate: bool);
+
+    /// Whether the aspect ratio is maintained on window resize.
+    fn maintain_aspect_ratio(&self) -> bool;
+    /// Set whether to `maintain` the aspect ratio on window resize.
+    fn set_maintain_aspect_ratio(&mut self, maintain: bool);
 }
+
 impl<T: MiniquadEventHandlerExt + 'static> UiService for MiniquadService<T> {
     fn capabilities(&self) -> UiCap {
         let image = Some(UiCapImage { rgb: true, ..Default::default() });
@@ -29,9 +41,6 @@ impl<T: MiniquadEventHandlerExt + 'static> UiService for MiniquadService<T> {
 }
 
 /// `miniquad`'s UI Service.
-///
-/// Works like a builder.
-#[derive(Default)]
 pub struct MiniquadService<T: MiniquadEventHandlerExt + 'static> {
     handler: Option<T>,
     conf: MiniquadConf,
@@ -39,33 +48,46 @@ pub struct MiniquadService<T: MiniquadEventHandlerExt + 'static> {
     /* pixel buffer settings */
     width: u32,
     height: u32,
-    linear_interpolation: bool,
-    maintain_aspect_ratio: bool,
 }
 
-/// # builder methods.
-impl<T: MiniquadEventHandlerExt + 'static> MiniquadService<T> {
-    /// Returns a new `MiniquadService` with the given `handler`, using default [`MiniquadConf`].
-    pub fn with(handler: T) -> Self {
+impl<T: MiniquadEventHandlerExt + 'static> Default for MiniquadService<T> {
+    /// Returns a default `MiniquadService` without a handler.
+    fn default() -> Self {
         Self {
-            handler: Some(handler),
+            handler: None,
             conf: MiniquadConf::default(),
             //
             width: 0,
             height: 0,
-            linear_interpolation: false,
-            maintain_aspect_ratio: false,
         }
     }
+}
+
+/// # builder methods
+impl<T: MiniquadEventHandlerExt + 'static> MiniquadService<T> {
+    /// Returns a default `MiniquadService` with the given `handler`,
+    pub fn with(handler: T) -> Self {
+        Self { handler: Some(handler), ..Default::default() }
+    }
+
     /// Initialize the miniquad rendering context, opening the window.
     ///
     /// This must be the last method to call.
+    ///
+    /// # Panics
+    /// Panics if the handler is not set (e.g. by constructing with `default()`)
     pub fn init(self) {
         let handler = self.handler.expect("Event handler is required");
         ::miniquad::start(self.conf, move || Box::new(handler.init()));
     }
 
     /* builder methods */
+
+    /// Sets the given handler.
+    pub fn handler(mut self, handler: T) -> Self {
+        self.handler = Some(handler);
+        self
+    }
 
     /// Sets the given miniquad configuration.
     pub fn conf(mut self, conf: MiniquadConf) -> Self {
@@ -84,8 +106,6 @@ impl<T: MiniquadEventHandlerExt + 'static> MiniquadService<T> {
     /* window */
 
     /// Sets the window title.
-    #[cfg(feature = "alloc")]
-    #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "alloc")))]
     pub fn title(mut self, title: impl ToString) -> Self {
         self.conf.window_title = title.to_string();
         self
@@ -124,14 +144,40 @@ impl<T: MiniquadEventHandlerExt + 'static> MiniquadService<T> {
         self.height = height;
         self
     }
-    /// If `linear == true` it uses [`FilterMode:Linear`], otherwise uses [`FilterMode::Nearest`].
-    pub fn linear_interpolation(mut self, linear: bool) -> Self {
-        self.linear_interpolation = linear;
+    /// If `linear == true` uses [`FilterMode:Linear`], otherwise [`FilterMode::Nearest`].
+    ///
+    /// If the `handler` has not been set, this does nothing.
+    pub fn interpolation(mut self, linear: bool) -> Self {
+        iif![let Some(h) = self.handler.as_mut(); h.set_interpolation(linear)];
         self
     }
-    /// Maintains the aspect ratio by drawing horizontal bands.
+
+    /// Set whether to `maintain` the aspect ratio on window resize.
+    ///
+    /// If the `handler` has not been set, this does nothing.
     pub fn maintain_aspect_ratio(mut self, maintain: bool) -> Self {
-        self.maintain_aspect_ratio = maintain;
+        iif![let Some(h) = self.handler.as_mut(); h.set_maintain_aspect_ratio(maintain)];
         self
+    }
+}
+
+/// # getters and setters
+impl<T: MiniquadEventHandlerExt + 'static> MiniquadService<T> {
+    /// Whether the aspect ratio is maintained on window resize.
+    pub fn get_interpolation(self) -> bool {
+        self.handler.as_ref().map(|h| h.interpolation()).unwrap_or_default()
+    }
+    /// Set whether to `maintain` the aspect ratio on window resize.
+    pub fn set_interpolation(&mut self, interpolate: bool) {
+        iif![let Some(h) = self.handler.as_mut(); h.set_interpolation(interpolate)];
+    }
+
+    /// Whether the aspect ratio is maintained on window resize.
+    pub fn get_maintain_aspect_ratio(self) -> bool {
+        self.handler.as_ref().map(|h| h.maintain_aspect_ratio()).unwrap_or_default()
+    }
+    /// Set whether to `maintain` the aspect ratio on window resize.
+    pub fn set_maintain_aspect_ratio(&mut self, maintain: bool) {
+        iif![let Some(h) = self.handler.as_mut(); h.set_maintain_aspect_ratio(maintain)];
     }
 }